Uninstall Doormat on migration (take 2)
