+++ src/brasil/gov/portal/dependencies.zcml
@@ -16,7 +16,7 @@
   <include package="plone.app.contenttypes" />
   <include package="plone.restapi" />
   <include package="Products.CMFPlone" />
-  <include package="Products.Doormat" />  <!-- TODO: remove -->
+  <include package="Products.Doormat" />  <!-- BBB: remove in 3.0 -->
   <include package="Products.PloneFormGen" />
   <include package="Products.PloneKeywordManager" />
   <include package="Products.RedirectionTool" />
