+++ src/brasil/gov/portal/upgrades/v10901/configure.zcml
@@ -52,11 +52,6 @@
         handler="..cook_javascript_resources"
         />
 
-    <genericsetup:upgradeStep
-        title="Uninstall Products.Doormat"
-        handler=".uninstall_doormat"
-        />
-
   </genericsetup:upgradeSteps>
 
 </configure>
